services/gpt-api/pyproject.toml
[project]
name = "gpt-api"
dynamic = ["version"]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.3",
    "syncit-core",
    "logger-lib",
    "environment-lib",
    "sentry-lib",
    "aiohttp>=3.13.2",
    "google-genai>=1.62.0",
]

[tool.uv.sources]
syncit-core = { workspace = true }
logger-lib = { workspace = true }
environment-lib = { workspace = true }
sentry-lib = { workspace = true }
